+++ src/forms/ERC20Form.sol
@@ -10,7 +10,7 @@ import {BaseForm} from "../BaseForm.sol";
 /// @notice Abstract implementation of Form for protocols using ERC20 vault shares.
 abstract contract ERC20Form is BaseForm {
     constructor(
-        uint256 chainId_,
+        uint80 chainId_,
         IStateRegistry stateRegistry_,
         ISuperFormFactory superFormFactory_
     ) BaseForm(chainId_, stateRegistry_, superFormFactory_) {}
